hal_record_write_begin: remove warning
hal/hal_ring.h: In function ‘int hal_record_write_begin(ringbuffer_t*, void**, size_t)’:
hal/hal_ring.h:151: warning: comparison between signed and unsigned integer expressions

since in
    free = (h->size + h->head - h->tail - 1) % h->size + 1; // -1 + 1 is needed for head==tail

the RHS operand of the modulo operation is guaranteed to be positive, it
is ok to declare free as 'size_t free'

+++ src/hal/hal_ring.h
@@ -139,7 +139,7 @@ static inline ring_size_t * _size_at(const ringbuffer_t *rb, const size_t off) /
  */
 static inline int hal_record_write_begin(ringbuffer_t *ring, void ** data, size_t sz)
 {
-    int free;
+    size_t free;
     ringheader_t *h = ring->header;
     size_t a = size_aligned(sz + sizeof(ring_size_t));
     if (a > h->size)
